[RWKV7] Add TileLang end-to-end backend: fullgraph training + cudagraph decode
New tilelang backend under fla/ops/rwkv7/backends/tilelang/ providing
custom_op-wrapped (graph-safe, fullgraph-compilable) ops:

- token_shift: tilelang replacement for the native token-shift module
  (fla/modules/token_shift.py gains a tilelang dispatch, cp variant too)
- fused_recurrent: decode-time fused recurrent step, cuda-graph friendly
- fused_addcmul / fused_k_update / gate_output_correction:
  elementwise fusions used by the layer
- kk_pre (fused l2norm K prep) and fused_gn_corr (fused group-norm +
  gate output correction): absorb the norm modules into tilelang kernels
- e2e: end-to-end layer path wiring the above into RWKV7Attention via
  FLA_RWKV7_TILELANG_E2E=1 (documented in ENVs.md)

Verified on 8xPRO6000 (local): 1-ulp parity vs native; training
max-autotune-no-cudagraphs rect 14587 tok/s (+2.1%), varlen 14253
(+3.4%), loss sequence exact vs baseline; decode under
mode="reduce-overhead" reaches real cudagraphs: 10.4x/7.6x/7.5x vs
eager at B=1/16/64 (g1h-1.5b), exact logits parity.

KNOWN ISSUE (marked in code/ENVs.md): with non-reentrant gradient
checkpointing + plain mode="max-autotune" (cudagraph trees enabled),
training hits "RuntimeError: Detected N tensor(s) in the cudagraph pool
not tracked as outputs". Root cause is torch-level: cudagraph-trees
routes warmup allocations into the private pool and its
check_memory_pool asserts every live pool block is a tracked graph
output; non-reentrant checkpointing stashes activations in
_CheckpointFrame.weak_holders which are live but not graph outputs.
Workaround: mode="max-autotune-no-cudagraphs" (default in our benches);
do NOT set triton.slow_path_cudagraph_asserts=False (risks silent
gradient corruption). A proper fix is follow-up work.

+++ fla/modules/l2warp.py
@@ -41,10 +41,10 @@ class L2Wrap(torch.autograd.Function):
     @staticmethod
     def backward(ctx, grad_output: torch.Tensor):
         maxx, ids = ctx.saved_tensors
-        glogits = torch.zeros(ctx.logits_shape, device=grad_output.device, dtype=grad_output.dtype)
+        glogits = torch.zeros(ctx.logits_shape, device=grad_output.device, dtype=maxx.dtype)
         # an autograd.Function must scale its input gradients by the upstream gradient; fold the
         # scalar grad_output into the sparse maxx to avoid a second full-size logits allocation
-        glogits.scatter_(-1, ids, maxx * grad_output)
+        glogits.scatter_(-1, ids, (maxx * grad_output).to(maxx.dtype))
         return grad_output, glogits, None
